EditButtonGroup - Menu compontent & cleanup
