+++ src/components/[guild]/EditButtonGroup/components/CustomizationButton/components/RemoveBackgroundImage.tsx
