+++ src/components/[guild]/EditButtonGroup/index.ts
@@ -1,0 +1,3 @@
+import EditButtonGroup from "./EditButtonGroup"
+
+export default EditButtonGroup
